requirements.txt
streamlit
pandas
numpy
plotly
seaborn
matplotlib.pyplot
